requirements.txt
strands-agents
strands-agents-tools
boto3
pytest
pytest-cov
hypothesis>=1.26.0
pytest>=7.0.0
pytest-cov>=4.0.0
hypothesis>=6.0.0